(refactor) Split up models.py into separate files pr. model.
Also changed to native UUID filed - which changes data representation in database.
