+++ members/models/waitinglist.py
@@ -1,0 +1,23 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+from django.db import models
+from django.utils import timezone
+
+
+class WaitingList(models.Model):
+    class Meta:
+        verbose_name="På venteliste"
+        verbose_name_plural='På ventelister'
+        ordering=['on_waiting_list_since']
+    person = models.ForeignKey('Person')
+    department = models.ForeignKey('Department')
+    on_waiting_list_since = models.DateField('Venteliste position', blank=False, null=False)
+    added_dtm = models.DateField('Tilføjet', blank=False, null=False, default=timezone.now)
+    def number_on_waiting_list(self):
+        return WaitingList.objects.filter(department = self.department, on_waiting_list_since__lt = self.on_waiting_list_since).count()+1
+    number_on_waiting_list.short_description = 'Position på venteliste'
+    def save(self, *args,**kwargs):
+        """ On creation set on_waiting_list """
+        if not self.id:
+            self.on_waiting_list_since = self.person.added
+        return super(WaitingList, self).save(*args, **kwargs)
