+++ members/templates/members/entry_page.html
@@ -82,5 +82,6 @@
           </div>
         </div>
       </div>
+    </div>
   </body>
 </html>
