+++ members/templates/members/volunteer_signup.html
@@ -95,5 +95,6 @@
           </div>
         </div>
       </div>
+    </div>
   </body>
 </html>
