+++ production_scripts/opt/backup_django_db.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-/usr/bin/sqlite3 /opt/database/db.sqlite3 ".dump" | bzip2 > /opt/database_backup/backup-$(date +%m-%d-%y-%H%M%S).sql.bz2
+/usr/bin/sqlite3 /opt/database/db.sqlite3 ".dump" | bzip2 > /opt/database_backup/backup-$(date +%y-%m-%d-%H%M%S).sql.bz2
 
 find /opt/database_backup -mtime +30 -exec rm {} \;
